Refactor data query tests
