--- data-query-api/checkstyle-suppressions.xml
@@ -1,14 +1,0 @@
-<!DOCTYPE suppressions PUBLIC
-  "-//Checkstyle//DTD SuppressionFilter Configuration 1.2//EN"
-  "https://checkstyle.org/dtds/suppressions_1_2.dtd">
-<suppressions>
-  <!--
-    The following resources have certain fields that are reported
-    as data absent reason extensions that is required to have a _
-    prefix by the Data Query specification.
-  -->
-  <suppress checks="MemberName" files=".*(/|\\)Immunization.java" message="Member name '(_status|_reported)'"/>
-  <suppress checks="MemberName" files=".*(/|\\)DiagnosticReport.java" message="Member name '(_performer)'"/>
-  <suppress checks="MemberName" files=".*(/|\\)MedicationOrder.java" message="Member name '(_prescriber)'"/>
-  <suppress checks="MemberName" files=".*(/|\\)Procedure.java" message="Member name '(_status)'"/>
-</suppressions>
