+++ data-query-tests/src/main/java/gov/va/api/health/sentinel/FhirTestClient.java
@@ -3,7 +3,7 @@ package gov.va.api.health.sentinel;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
-import gov.va.api.health.dataquery.api.resources.OperationOutcome;
+import gov.va.api.health.dstu2.api.resources.OperationOutcome;
 import io.restassured.http.Method;
 import io.restassured.response.Response;
 import io.restassured.response.ResponseBody;
