+++ data-query/pom.xml
@@ -29,11 +29,6 @@
       <artifactId>service-auto-config</artifactId>
     </dependency>
     <dependency>
-      <groupId>gov.va.api.health</groupId>
-      <artifactId>data-query-api</artifactId>
-      <version>${project.version}</version>
-    </dependency>
-    <dependency>
       <groupId>gov.va.dvp</groupId>
       <artifactId>cdw-schemas</artifactId>
     </dependency>
@@ -42,38 +37,89 @@
       <artifactId>rest-assured</artifactId>
     </dependency>
     <dependency>
-      <groupId>gov.va.api.health</groupId>
-      <artifactId>data-query-api</artifactId>
-      <version>${project.version}</version>
-      <classifier>tests</classifier>
+      <groupId>org.glassfish</groupId>
+      <artifactId>javax.el</artifactId>
       <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>gov.va.api.health</groupId>
+      <artifactId>ids-client</artifactId>
+      <version>1.1.2</version>
+    </dependency>
+    <dependency>
+      <groupId>org.reflections</groupId>
+      <artifactId>reflections</artifactId>
+      <version>${reflections.version}</version>
+      <scope>test</scope>
+    </dependency>
+    <dependency>
+      <groupId>org.springframework</groupId>
+      <artifactId>spring-beans</artifactId>
+      <scope>compile</scope>
     </dependency>
   </dependencies>
   <build>
     <plugins>
       <plugin>
-        <artifactId>maven-dependency-plugin</artifactId>
+        <groupId>io.swagger.core.v3</groupId>
+        <artifactId>swagger-maven-plugin</artifactId>
         <executions>
           <execution>
-            <id>swagger</id>
-            <phase>generate-resources</phase>
+            <id>dataquery-swag</id>
+            <phase>compile</phase>
             <goals>
-              <goal>copy</goal>
+              <goal>resolve</goal>
             </goals>
             <configuration>
-              <artifactItems>
-                <artifactItem>
-                  <groupId>${project.groupId}</groupId>
-                  <artifactId>data-query-api</artifactId>
-                  <version>${project.version}</version>
-                  <type>yaml</type>
-                  <classifier>openapi</classifier>
-                  <overWrite>true</overWrite>
-                  <outputDirectory>${project.build.directory}/swagger</outputDirectory>
-                  <destFileName>openapi.yaml</destFileName>
-                </artifactItem>
-              </artifactItems>
+              <resourceClasses>
+                <resourceClass>gov.va.api.health.dataquery.service.api.DataQueryService</resourceClass>
+              </resourceClasses>
+              <outputFileName>openapi</outputFileName>
+              <outputPath>${project.build.outputDirectory}</outputPath>
+              <outputFormat>JSONANDYAML</outputFormat>
+              <prettyPrint>true</prettyPrint>
             </configuration>
+          </execution>
+        </executions>
+      </plugin>
+      <plugin>
+        <artifactId>maven-antrun-plugin</artifactId>
+        <executions>
+          <execution>
+            <phase>compile</phase>
+            <configuration>
+              <tasks>
+                <!--
+                  Sadness here.
+                  Swagger is ignore @JsonProperty annotations on enums. This big bit of sadness changes generated swagger _ to - in enums.
+                -->
+                <replaceregexp byline="true" file="${project.build.outputDirectory}/openapi.yaml" flags="g" match="([a-z])_([a-z])" replace="\1-\2"/>
+                <replaceregexp byline="true" file="${project.build.outputDirectory}/openapi.yaml" flags="g" match="- _([a-z])" replace="- \1"/>
+                <replaceregexp byline="true" file="${project.build.outputDirectory}/openapi.json" flags="g" match="([a-z])_([a-z])" replace="\1-\2"/>
+                <replaceregexp byline="true" file="${project.build.outputDirectory}/openapi.json" flags="g" match="(,|\[) &quot;_([a-z])" replace="\1 &quot;\2"/>
+                <!--
+                  This group is because the swagger auto-generated for dates, times, and datetimes examples doesn't play nice with how swagger is actually consumed.
+                  The first regex captures datetimes and times. The second one covers left over dates.
+                -->
+                <replaceregexp byline="true" file="${project.build.outputDirectory}/openapi.yaml" flags="g" match="\b(([0-9]{4}(-(0[1-9]|1[0-2])(-(0[0-9]|[1-2][0-9]|3[0-1]))?)T)?(([01][0-9]|2[0-3]):[0-5][0-9]:[0-5][0-9](\.[0-9]+)?)(Z|(\+|-)((0[0-9]|1[0-3]):[0-5][0-9]|14:00))?)\b" replace="'\1'"/>
+                <replaceregexp byline="true" file="${project.build.outputDirectory}/openapi.yaml" flags="g" match="( ([0-9]{4}-(0[1-9]|1[0-2])(-(0[0-9]|[1-2][0-9]|3[0-1]))?))" replace=" '\2'"/>
+              </tasks>
+            </configuration>
+            <goals>
+              <goal>run</goal>
+            </goals>
+          </execution>
+        </executions>
+      </plugin>
+      <plugin>
+        <artifactId>maven-jar-plugin</artifactId>
+        <executions>
+          <execution>
+            <id>test-jar</id>
+            <phase>package</phase>
+            <goals>
+              <goal>test-jar</goal>
+            </goals>
           </execution>
         </executions>
       </plugin>
@@ -81,20 +127,6 @@
         <groupId>org.codehaus.mojo</groupId>
         <artifactId>build-helper-maven-plugin</artifactId>
         <executions>
-          <execution>
-            <id>add-some-swagger</id>
-            <phase>generate-resources</phase>
-            <goals>
-              <goal>add-resource</goal>
-            </goals>
-            <configuration>
-              <resources>
-                <resource>
-                  <directory>${project.build.directory}/swagger</directory>
-                </resource>
-              </resources>
-            </configuration>
-          </execution>
           <execution>
             <id>conformance-timestamp</id>
             <goals>
